requirements.txt
# Install the package in editable mode first:
#   pip install -e .
# Then install dev/test dependencies:
fastapi
uvicorn
mcp
pytest>=8.0
pytest-asyncio>=0.23
fakeredis[aioredis]>=2.20
httpx>=0.27
fastapi-key-auth>=0.9.0